实现了部分添加删除分类的功能
