+++ NewsApp/app/src/main/res/values/dimens.xml
@@ -5,4 +5,5 @@
     <dimen name="item_width">200dp</dimen>
     <dimen name="text_margin">16dp</dimen>
     <dimen name="category_height">35dp</dimen>
+    <dimen name="category_item_height">100dp</dimen>
 </resources>
